+++ packages/expo-modules-core/ios/Core/Protocols/AnyExpoView.swift
@@ -1,0 +1,5 @@
+public protocol AnyExpoView: RCTView {
+  var appContext: AppContext? { get }
+
+  init(appContext: AppContext?)
+}